% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snippets.R
\name{add_snippets}
\alias{add_snippets}
\title{Add useful snippets}
\usage{
add_snippets()
}
\value{
Nothing. Creates or modifies a file with custom snippets.
}
\description{
This function creates or edits an r.snippets file with custom snippets. After
running this function you will have functions for creating a connection
object for databases and for setting common options.
}
\examples{
# Not run
if (FALSE) add_snippets()
}
