% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snippets.R
\name{custom_snippet_file_exsists}
\alias{custom_snippet_file_exsists}
\title{Check if custom snippet file already exists}
\usage{
custom_snippet_file_exsists()
}
\value{
Boolean, TRUE if a custom snipept file exists, else FALSE
}
\description{
Check if custom snippet file already exists
}
\examples{
custom_snippet_file_exsists()
}
