% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snippets.R
\name{find_snippet_path}
\alias{find_snippet_path}
\title{Find path to custom snippet file}
\usage{
find_snippet_path()
}
\value{
Returns the path to the custom snippet file as a string
}
\description{
Find path to custom snippet file
}
\examples{
find_snippet_path()

}
